Fix cleanup, factor out getJsonMetadata
